#136 fix: move to services layer
